Add files via upload
A Simple Website Where You can Give  and Take MCQ based Quiz Exam.

Just You Have to Add Your Question , Options and , the Correct Answer.

There is  Timer of 10sec for Each Question .With in This Time the User Have to mark there Options Otherwise the Next Question  will be displayed. You Can Change the Time for Each Question Also. 

At the End of the Exam Your Score Will be displayed along with the Result Summary.